pdf-parser: Add a recursion limit for object nesting
