+++ crates/pdf-content-stream/src/pdf_operator/variants.rs
@@ -1,4 +1,4 @@
-use pdf_parser::{PdfParser, traits::CommentParser};
+use pdf_parser::{parser::PdfParser, traits::CommentParser};
 use pdf_tokenizer::PdfToken;
 
 use crate::{
